backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.0.0
requests>=2.31.0
httpx>=0.25.0
openai>=1.52.0
google-cloud-storage>=2.18.0

PyPDF2==3.0.1
openpyxl==3.1.2
python-multipart==0.0.9
